+++ packages/core/src/loop/types.ts
@@ -1,4 +1,3 @@
-import type { WritableStream } from 'node:stream/web';
 import type { SharedV2ProviderOptions } from '@ai-sdk/provider-v5';
 import type { CallSettings, IdGenerator, StopCondition, ToolChoice, ToolSet, StepResult, ModelMessage } from 'ai-v5';
 import z from 'zod';
@@ -21,6 +20,7 @@ import type {
   ModelManagerModelConfig,
 } from '../stream/types';
 import type { MastraIdGenerator } from '../types';
+import type { OutputWriter } from '../workflows/types';
 
 export type StreamInternal = {
   now?: () => number;
@@ -96,6 +96,7 @@ export type LoopOptions<Tools extends ToolSet = ToolSet, OUTPUT extends OutputSc
   requireToolApproval?: boolean;
   agentId: string;
   toolCallConcurrency?: number;
+  agentName?: string;
   requestContext?: RequestContext;
   methodType: ModelMethodType;
 };
@@ -119,7 +120,7 @@ export type LoopRun<Tools extends ToolSet = ToolSet, OUTPUT extends OutputSchema
 export type OuterLLMRun<Tools extends ToolSet = ToolSet, OUTPUT extends OutputSchema = undefined> = {
   messageId: string;
   controller: ReadableStreamDefaultController<ChunkType<OUTPUT>>;
-  writer: WritableStream<ChunkType<OUTPUT>>;
+  outputWriter: OutputWriter;
 } & LoopRun<Tools, OUTPUT>;
 
 export const PRIMITIVE_TYPES = z.enum(['agent', 'workflow', 'none', 'tool']);