+++ packages/playground-ui/src/domains/workflows/runs/workflow-run-list.tsx
@@ -2,14 +2,16 @@ import { Txt } from '@/ds/components/Txt';
 import { Icon } from '@/ds/icons';
 import { WorkflowRunStatus } from '@mastra/core/workflows';
 import { Check, CirclePause, CircleSlash, Clock, Plus, X } from 'lucide-react';
-import { useWorkflowRuns } from '@/hooks/use-workflow-runs';
-import { ThreadItem, ThreadLink, ThreadList, Threads } from '@/components/threads';
+import { useDeleteWorkflowRun, useWorkflowRuns } from '@/hooks/use-workflow-runs';
+import { ThreadDeleteButton, ThreadItem, ThreadLink, ThreadList, Threads } from '@/components/threads';
 import { useLinkComponent } from '@/lib/framework';
 import { formatDate } from 'date-fns';
 import { Skeleton } from '@/components/ui/skeleton';
 import { Badge } from '@/ds/components/Badge';
 import Spinner from '@/components/ui/spinner';
 import { useInView } from '@/hooks';
+import { AlertDialog } from '@/components/ui/alert-dialog';
+import { useState } from 'react';
 
 export interface WorkflowRunListProps {
   workflowId: string;
@@ -17,8 +19,21 @@ export interface WorkflowRunListProps {
 }
 
 export const WorkflowRunList = ({ workflowId, runId }: WorkflowRunListProps) => {
-  const { Link, paths } = useLinkComponent();
+  const [deleteRunId, setDeleteRunId] = useState<string | null>(null);
+
+  const { Link, paths, navigate } = useLinkComponent();
   const { isLoading, data: runs, setEndOfListElement, isFetchingNextPage } = useWorkflowRuns(workflowId);
+  const { mutateAsync: deleteRun } = useDeleteWorkflowRun(workflowId);
+
+  const handleDelete = async (runId: string) => {
+    try {
+      await deleteRun({ runId });
+      setDeleteRunId(null);
+      navigate(paths.workflowLink(workflowId));
+    } catch (_error) {
+      setDeleteRunId(null);
+    }
+  };
 
   const actualRuns = runs || [];
 
@@ -59,7 +74,7 @@ export const WorkflowRunList = ({ workflowId, runId }: WorkflowRunListProps) => 
                     <WorkflowRunStatusBadge status={run.snapshot.status} />
                   </div>
                 )}
-                <span className="truncate max-w-[14rem] text-muted-foreground">{run.runId}</span>
+                <span className="truncate max-w-32 text-muted-foreground">{run.runId}</span>
                 <span>
                   {typeof run?.snapshot === 'string'
                     ? ''
@@ -68,10 +83,22 @@ export const WorkflowRunList = ({ workflowId, runId }: WorkflowRunListProps) => 
                       : ''}
                 </span>
               </ThreadLink>
+
+              <ThreadDeleteButton onClick={() => setDeleteRunId(run.runId)} />
             </ThreadItem>
           ))}
         </ThreadList>
       </Threads>
+
+      <DeleteRunDialog
+        open={!!deleteRunId}
+        onOpenChange={() => setDeleteRunId(null)}
+        onDelete={() => {
+          if (deleteRunId) {
+            handleDelete(deleteRunId);
+          }
+        }}
+      />
 
       {isFetchingNextPage && (
         <div className="flex justify-center items-center">
@@ -139,4 +166,28 @@ const WorkflowRunStatusBadge = ({ status }: WorkflowRunStatusProps) => {
   }
 
   return <Badge variant="default">{status}</Badge>;
+};
+
+interface DeleteRunDialogProps {
+  open: boolean;
+  onOpenChange: (n: boolean) => void;
+  onDelete: () => void;
+}
+const DeleteRunDialog = ({ open, onOpenChange, onDelete }: DeleteRunDialogProps) => {
+  return (
+    <AlertDialog open={open} onOpenChange={onOpenChange}>
+      <AlertDialog.Content>
+        <AlertDialog.Header>
+          <AlertDialog.Title>Are you absolutely sure?</AlertDialog.Title>
+          <AlertDialog.Description>
+            This action cannot be undone. This will permanently delete the workflow run and remove it from our servers.
+          </AlertDialog.Description>
+        </AlertDialog.Header>
+        <AlertDialog.Footer>
+          <AlertDialog.Cancel>Cancel</AlertDialog.Cancel>
+          <AlertDialog.Action onClick={onDelete}>Continue</AlertDialog.Action>
+        </AlertDialog.Footer>
+      </AlertDialog.Content>
+    </AlertDialog>
+  );
 };